backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-elasticsearch-dsl==8.0
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
drf-social-oauth2==3.1.0
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.0
gunicorn==22.0.0
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
